Implement coloring functions
On grid creation, two even listeners are added to the document, a mousedown listener and a mouseup listener. On mousedown, if the event target is a square in the grid, the square will be colored black, and an event listener is added to the grid for mouseover of children elements that will color the square. On mousedown, the mouseover event listener is removed from the grid.